standardised simple model usage
